move get_pxd_path to setup_utils
